+++ backend/src/main/java/com/carffeine/carffeine/station/controller/auth/MemberController.java
@@ -1,6 +1,6 @@
 package com.carffeine.carffeine.station.controller.auth;
 
-import com.carffeine.carffeine.station.service.auth.UserService;
+import com.carffeine.carffeine.station.service.auth.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api")
 @RequiredArgsConstructor
-public class UserController {
+public class MemberController {
 
-    private final UserService userService;
+    private final MemberService memberService;
 
     @PostMapping("/jwt")
     public ResponseEntity<String> createJwt() {
