+++ backend/src/main/java/com/carffeine/carffeine/station/controller/auth/SearchController.java
@@ -1,7 +1,7 @@
 package com.carffeine.carffeine.station.controller.auth;
 
 import com.carffeine.carffeine.station.controller.auth.dto.MemberResponse;
-import com.carffeine.carffeine.station.domain.user.User;
+import com.carffeine.carffeine.station.domain.member.Member;
 import com.carffeine.carffeine.station.infrastructure.api.AuthMember;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/search")
 public class SearchController {
     @PostMapping
-    public ResponseEntity<MemberResponse> startSearching(@AuthMember User user) {
-        MemberResponse response = new MemberResponse(user.getId(), user.getEmail());
+    public ResponseEntity<MemberResponse> startSearching(@AuthMember Member member) {
+        MemberResponse response = new MemberResponse(member.getId(), member.getEmail());
         return ResponseEntity.ok().body(response);
     }
 }
