+++ backend/src/main/java/com/carffeine/carffeine/station/domain/filter/AuthenticateMember.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class AuthenticateUser {
+public class AuthenticateMember {
     private Long id;
     private String email;
 }
