+++ backend/src/main/java/com/carffeine/carffeine/station/domain/member/MemberRepository.java
@@ -1,0 +1,8 @@
+package com.carffeine.carffeine.station.domain.member;
+
+import org.springframework.data.repository.Repository;
+
+public interface MemberRepository extends Repository<Member, Long> {
+
+    Member findById(Long id);
+}
