--- backend/src/main/java/com/carffeine/carffeine/station/domain/user/UserRepository.java
@@ -1,8 +1,0 @@
-package com.carffeine.carffeine.station.domain.user;
-
-import org.springframework.data.repository.Repository;
-
-public interface UserRepository extends Repository<User, Long> {
-
-    User findById(Long id);
-}
