--- backend/src/main/java/com/carffeine/carffeine/station/service/user/dto/UserLoginRequest.java
@@ -1,4 +1,0 @@
-package com.carffeine.carffeine.station.service.user.dto;
-
-public record UserLoginRequest(String email, String password) {
-}
