+++ backend/src/test/java/com/carffeine/carffeine/station/MockBeanInjection.java
@@ -1,9 +1,9 @@
 package com.carffeine.carffeine.station;
 
 import com.carffeine.carffeine.station.domain.jwt.JwtProvider;
-import com.carffeine.carffeine.station.domain.user.UserRepository;
+import com.carffeine.carffeine.station.domain.member.MemberRepository;
 import com.carffeine.carffeine.station.infrastructure.api.AuthMemberResolver;
-import com.carffeine.carffeine.station.service.auth.UserService;
+import com.carffeine.carffeine.station.service.auth.MemberService;
 import com.carffeine.carffeine.station.service.congestion.CongestionService;
 import com.carffeine.carffeine.station.service.station.StationService;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -16,9 +16,9 @@ public class MockBeanInjection {
     @MockBean
     protected StationService stationService;
     @MockBean
-    protected UserRepository userRepository;
+    protected MemberRepository memberRepository;
     @MockBean
-    protected UserService userService;
+    protected MemberService memberService;
     @MockBean
     protected CongestionService congestionService;
 }
